Sections, Klasses, Lessons templates created. ModalBody is in Form component wrapping form. ModalFooter is in Form too. Contains submit button. Submit Button gets values from props and passes to handleSubmit when clicked
